requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
python-pptx>=0.6.21
streamlit>=1.28.0
pytest>=7.4.0
openai>=1.3.0
groq>=0.4.0
requests>=2.31.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
reportlab>=4.0.0
Pillow>=10.0.0
sqlalchemy>=2.0.0

